refactor(pxi): extract LLM-evaluator draft registration into a hook
Pull the inline draft-host registration effect out of
EditLLMEvaluatorDialogContent into a dedicated
useLlmEvaluatorDraftRegistration hook, addressing review feedback that the
component held too much inline logic. The hook is colocated with the
component to keep the agent/tools/llmEvaluatorDraft module React/store-free.

Also drop the lore-internal "Load-bearing" scaffolding prefix from an
agents.py comment while keeping the cross-function-contract fact it carried.
